server/requirements.txt
# Ядро langchain — одно поколение, langchain-core 1.x
langchain==1.2.10
langchain-community>=0.4.0
langchain-text-splitters>=0.4.0
langchain-openai==1.1.0
langchain-google-genai==3.0.0
langchain-weaviate==0.0.6

# Загрузка документов (PDF, DOCX)
pypdf>=4.0.0
docx2txt>=0.8

# Веб-сервер
fastapi==0.115.12
uvicorn==0.34.0

# Валидация
pydantic==2.10.6
pydantic-settings>=2.10.1,<3.0.0

# База данных
psycopg2-binary==2.9.11
weaviate-client>=4.10.0,<5.0.0

# Утилиты
loguru==0.7.3
python-multipart
nltk
natasha
langgraph>=0.2.0
aiohttp
bs4